+++ src/chunk/mod.rs
@@ -9,6 +9,7 @@ pub(crate) mod chunk_forward_tsn;
 pub(crate) mod chunk_header;
 pub(crate) mod chunk_heartbeat;
 pub(crate) mod chunk_heartbeat_ack;
+pub(crate) mod chunk_i_forward_tsn;
 pub(crate) mod chunk_init;
 pub mod chunk_payload_data;
 pub(crate) mod chunk_reconfig;